### 2016-12-07 #133 @whedon
[REVIEW]: PETRARCH2: Another Event Coding Program
**Submitting author:** @johnb30 (<a href="http://orcid.org/0000-0001-7811-4399">John Beieler</a>)
**Repository:** <a href="https://github.com/openeventdata/petrarch2" target ="_blank">https://github.com/openeventdata/petrarch2</a>
**Version:** v1.0.0
