### 2016-12-24 openeventdata/petrarch2#37 @RichardLitt
Install instructions reference incorrect petrarch version
The install instructions currently call for:

```
pip install git+https://github.com/openeventdata/petrarch2.git
```

I was able to get this working with:

```
pip install git+https://github.com/openeventdata/petrarch2.git --ignore-installed
```

As I am running El Capitan, which has some issues. Unfortunately, the following steps didn't work:

```
petrach -h
```

Because this installs the executable:

```
petrarch2
```

The instructions should be modified to add the digit `2` to the program name.

This was brought up in https://github.com/openjournals/joss-reviews/issues/133.
