### 2016-12-24 openeventdata/petrarch2#38 @RichardLitt
Add a Contribute section to README
The README does not have a contribution section, which makes it hard for me to know if I can open PRs, open issues, or what pathways there are for contribution in general. While there is a [Contribute section in the docs](https://petrarch2.readthedocs.io/en/latest/contributing.html), this only covers coding guidelines and a task-like process, not whether or not this is encouraged. It would be good to make this explicit here, in the README.

This proposed change is blocking https://github.com/openjournals/joss-reviews/issues/133.
