### 2016-12-24 #133 @RichardLitt
I believe that the software and possibly also the test dataset need DOIs. That, and the issues above, are blocking this. Not major blockers, just some small reworking needed.
