### 2016-12-24 #133 @RichardLitt
Not sure how to say I consider my first pass done; given the linked issues above, this can be closed. Throwing this back into the author's camp for now.
