### 2016-12-24 #133 @johnb30
Hi @RichardLitt, thanks a ton for the review! We'll work over the next few days to update things and will drop a note in this issue when things are updated.
