### 2016-12-24 #133 @RichardLitt
Sounds good.

General comments: This is well done. On a note that isn't technically needed but would be great - I'd like the answer to "Do the authors clearly state what problems the software is designed to solve and who the target audience is?" to be clearer in the document. Yes, you cover this, but I think the audience could be explicitly stated. I understand where the work is coming from, but some readers won't.

Otherwise, LGTM.
