### 2016-12-27 03e3b7d
Update README install instructions.
