### 2016-12-27 19dbd0b
Update install instructions in docs.
