### 2016-12-27 096595e
Update contributing guidelines in the docs.
