### 2016-12-27 c6613d7
More to the docs.
