### 2016-12-27 openeventdata/petrarch2#38 @johnb30
Addressed in #40.
