### 2016-12-27 a78c6c3
Update JOSS paper.
