### 2016-12-27 openeventdata/petrarch2#37 @RichardLitt
Looks good to me. Close?
