### 2016-12-27 openeventdata/petrarch2#37 @johnb30
👍

Thanks again for your feedback!
