### 2016-12-27 #133 @johnb30
@RichardLitt as you saw I updated things in those two issues. I also added more into the paper body itself about the target audience and use cases. We greatly appreciate all of your feedback and input.

The only thing I wasn't sure about was the DOIs. If it's just the DOI for the paper I think that gets generated when the paper is accepted. Maybe @arfon can chime in?
