### 2016-12-27 cf5f47f
Add JOSS badge.
