### 2016-12-27 #133 @RichardLitt
Your changes also look good to me. @arfon can you clarify about the DOIs?
