### 2016-12-30 #133 @arfon
> The only thing I wasn't sure about was the DOIs. If it's just the DOI for the paper I think that gets generated when the paper is accepted. Maybe @arfon can chime in?

@johnb30 - we need you to create an archive of the repository with Zenodo or figshare - see [this guide](https://guides.github.com/activities/citable-code/) for example. When you have done this, please let us know the archive DOI here.

We will then issue a DOI for the JOSS paper separately.
