### 2016-12-30 #133 @johnb30
I just minted a DOI. The number is (I think...) `10.5281/zenodo.224637` and the archive is at https://zenodo.org/record/224637#.WGXGQ7YrJE4. Let me know if anything else is needed from my end.
